requirements.txt
# requirements.txt
# Core runtime (exact versions from your environment)
streamlit==1.51.0
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-google-genai==3.0.2
langchain-groq==1.0.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
qdrant-client==1.15.1
groq==0.33.0
tiktoken==0.12.0
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5
pypdf==6.2.0
docx2txt
pandas==2.3.3
numpy==2.3.4
PyYAML==6.0.3

# Testing & QA (pinned stable versions)
pytest==8.3.2
pytest-mock==3.14.0
pytest-cov==5.1.0
responses==0.25.0
vcrpy==5.1.0

black==24.8.0
ruff==0.6.0
flake8==7.1.0
pre-commit==3.8.0